Get the UI working
For some reason you had two routes, one for `canvas` and one for `index`, both
of which had practically the same ERB.

This adds extra parts which are not useful.

You also had broken links to the javascript files; which would result in the
application not working.

I strongly recommend before merging pull requests you pull down the changes and
run them on a machine other than the committers to ensure that it has all the
parts necessary.